Image file name correction
Fix issue where images were not appearing for Sample Report Col AG, Sample Relationship Col AE, Protocol Step Col AG and HO, Protocol relationship Col AH.  The reason for this is because the "Col" in the image file name in the qmd file was not capitalized, while the actual file name had it capitalized.  To fix this, the qmd file was edited.  It should be noted that capitalization does not matter in Rstudio.